Indirect SIMD implementation to 128 bit
This moves the implementation into a simpler one
relying on 128 bit vectors rather than what was
here prior, with 256 using the control vector of
indicies to drive shuffling (permute, actually).

This is slower, but it is a bit clearer to
understand, and I've updated the docs to match.

Given the direct implementation is *far* faster
I am not worried about optimising this particular
approach.